requirements.txt
async-timeout==4.0.2
autopep8==1.7.0
boto3==1.24.79
botocore==1.27.79
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
docformatter==1.5.0
idna==3.4
jmespath==1.0.1
packaging==21.3
pep8radius==0.9.2
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1
rq==1.11.0
s3transfer==0.6.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
untokenize==0.1.1
urllib3==1.26.12
wrapt==1.14.1
yapf==0.32.0
